migrate device model into own domain
