+++ domain/avd/avd.go
@@ -10,3 +10,6 @@ type AvdDeviceConfig struct {
 	// Overlay
 	// Platform
 }
+
+
+func
